Allow using a cookie as a transport for the token
A new configuration 'JWT_AUTH_COOKIE' has been added. If it set
to None (the default), everything works the same.

It can be set to a string to use as the name of the cookie that
can be used to carry the token. Note the 'Authorization' header is
still accepted as the primary method, but it falls back to the
designated cookie if present.

When a token is requested and 'JWT_AUTH_COOKIE' is set, the response
will set the given cookie with the token string.

Fix #120