Allow a setting to omit the user_id from the token payload.
The default value preserves the current behaviour.

As far as I can tell, the default behaviour doesn't use this value to look up the user, so omitting it should generally be safe (unless an application is relying on it when using the token).

In my particular case, this primary key is a serial number, which I would prefer not to leak to the outside world.