Use a demo project for running the test suite
When all test environments were run locally, the migration files would
persist. This would result in `NodeNotFoundError` being raised when
ran against a version of Django that does not contain that migration
file.

This change introduces a migration file that is connected to Django's
0008 `auth` migration. Since the change to the custom `User` model is
trivial, this should not be an issue when tested against future versions
of Django.